[processor] Refactor processing logic out of CGI
Keep the server (main.py) simple and move business logic related to
processing tasks into a separate module.

Also add more logging.
